pyproject.toml
[tool]
[tool.poetry]
name = "gitlab4"
version = "0.1.0"
homepage = "https://github.com/luiscberrocal/gitlab4"
description = "Gitlab API wrapper with Pydantic."
authors = ["Luis C. Berrocal <luis.berrocal.1942@gmail.com>"]
readme = "README.md"
license =  "MIT"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
packages = [
    { include = "gitlab4" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
click = "^8.1.7"
rich = "^13.6.0"
pydantic = "^2.5.3"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
flake8-docstrings = "^1.7.0"
factory-boy = "^3.3.0"
pytz = "^2023.3"
bump2version = "^1.0.1"
twine = "^4.0.2"
coverage = "^7.3.2"
tox = "^4.11.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
freezegun = "^1.2.2"
python-dotenv = "^1.0.0"


[tool.poetry.group.doc.dependencies]
mkdocs = "^1.5.3"
mkdocs-include-markdown-plugin = "^6.0.3"
mkdocs-material = "^9.4.6"
mkdocstrings = "^0.23.0"
mkdocs-autorefs = "^0.5.0"

[tool.poetry.scripts]
gitlab4 = 'gitlab4.cli:main'
[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 6
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = false
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
